with-temp-gopath.sh: Use separate variable for temp GOPATH
Instead of directly overriding GOPATH and worrying about whether it may
point to an unintended location before calling mktemp, just use a
separate variable that's set to any empty string before registering the
trap. This closes the tiny race between calling mktemp and registering
the trap afterwards.

https://phabricator.endlessm.com/T25314

+++ with-temp-gopath.sh
@@ -18,14 +18,18 @@ if [[ $# -eq 0 ]]; then
   exit 64  # EX_USAGE from sysexits.h
 fi
 
+TEMP_GOPATH=
 cleanup() {
-  rm -rf "$GOPATH"
+  if [[ -n "$TEMP_GOPATH" && -d "$TEMP_GOPATH" ]]; then
+    rm -rf "$GOPATH"
+  fi
 }
-GOPATH=$(mktemp -d)
 trap cleanup EXIT
+TEMP_GOPATH=$(mktemp -d)
+GOPATH="$TEMP_GOPATH"
 export GOPATH
 
-GOPATH_SRCDIR="$GOPATH/src/$PACKAGE"
+GOPATH_SRCDIR="$TEMP_GOPATH/src/$PACKAGE"
 SRCDIR=$(readlink -f "$PWD")
 mkdir -p "$(dirname "$GOPATH_SRCDIR")"
 ln -s "$SRCDIR" "$GOPATH_SRCDIR"
